pyproject.toml
[tool.poetry]
name = "aqbt"
version = "0.0.1"
description = "Aquarium strain builder tools"
authors = ["Justin Vrana <justin.vrana@gmail.com>"]
readme = 'README.md'

[tool.poetry.dependencies]
python = "^3.6"
biopython = "^1.73"
networkx = "^2.4"
tqdm = "^4.36"
primer3plus = "^1.0"
pydent = "^0.1.5a14"
loggable-jdv = "^0.1.7"
benchlingapi = "^2.1.11"
colour = "^0.1.5"
validators = "^0.14.1"
bcbio-gff = "^0.6.6"
pyblastbio = "^0.6.3"
google-api-python-client = "^1.7"
google-auth-httplib2 = "^0.0.3"
google-auth-oauthlib = "^0.4.1"
toml = "^0.10.0"
dasi = "^0.0.20"


[tool.poetry.dev-dependencies]
black = { version = "^19.3b0", python = "^3.6" }
pytest = "^5.2"
pre-commit = "^1.18"
toml = "^0.10.0"
dictdiffer = "^0.8.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
